Added a sneaky sentence to README.
In particular, I wanted to generate a commit for some project
to which I could play with a long commit message. Apparently,
they, long commit messages that is, are not uncommon. It does
make some sense: There is the punchy headline as the first
sentence, and a more detailed description down below, but what
I would like to know is how this will look on Github.